pyproject.toml
[project]
name = "bug-clarity-evaluator"
version = "0.1.0"
description = "A bug clarity evaluator pipeline for bugzilla instances"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiodns>=4.0.0",
    "aiohttp>=3.13.3",
    "catboost>=1.2.10",
    "lightgbm>=4.6.0",
    "numpy>=2.4.2",
    "polars>=1.38.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.3",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "xgboost-cpu>=3.2.0",
]

[tool.ty.environment]
root = ["./src/build_dataset", "./src/build_model"]

[tool.uv.sources]
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[dependency-groups]
dev = [
    "ruff>=0.15.2",
    "ty>=0.0.18",
]
